+++ deploy/assets/js/distortion-grid.js
@@ -3,7 +3,7 @@
 // Version: v1.790
 
 (function() {
-console.log('[DistortionGrid] v3.043 Loaded'); 
+console.log('[DistortionGrid] v3.044 Loaded'); 
 
 class DistortionGrid {
     constructor(parentElement, index) {