+++ deploy/assets/js/scroll-flipper.js
@@ -8,7 +8,7 @@
  */
 
 (function() {
-    console.log('[ScrollFlipper v3.043] Loading High-Performance Mode...');
+    console.log('[ScrollFlipper v3.044] Loading High-Performance Mode...');
 
     let isRunning = false;
     let track, stickyContainer, cards, triggers, cardParent;